fix: fuacet as component, faucet service fix: cleanup token component
