+++ packages/boba/gateway/src/containers/wallet/token/Token.js
@@ -3,7 +3,7 @@ import { useDispatch, useSelector } from 'react-redux'
 
 import { selectlayer1Balance, selectlayer2Balance } from 'selectors/balanceSelector'
 import { selectLoading } from 'selectors/loadingSelector'
-import { selectAccountEnabled, selectLayer, selectWalletAddress } from 'selectors/setupSelector'
+import { selectAccountEnabled, selectLayer } from 'selectors/setupSelector'
 import { selectNetwork } from 'selectors/networkSelector'
 import { selectTokens } from 'selectors/tokenSelector'
 import { selectTransactions } from 'selectors/transactionSelector'
@@ -11,28 +11,25 @@ import { selectTransactions } from 'selectors/transactionSelector'
 import { fetchLookUpPrice } from 'actions/networkAction'
 import { setActiveHistoryTab } from 'actions/uiAction'
 
-import { openAlert } from 'actions/uiAction'
-
 import * as S from './Token.styles'
 import * as G from '../../Global.styles'
 
-import twitter from 'images/twitter.png'
 
-import { Box, Typography, CircularProgress, Input } from '@mui/material'
+import { Box, Typography, CircularProgress } from '@mui/material'
 import { tokenTableHeads } from './token.tableHeads'
 
 import ListToken from 'components/listToken/listToken'
 import Button from 'components/button/Button'
 import Link from 'components/icons/LinkIcon'
 import Pulse from 'components/pulse/PulsingBadge'
-import Copy from 'components/copy/Copy'
 
 import { isEqual, orderBy } from 'lodash'
 
 import networkService from 'services/networkService'
 
-import { Md5 } from "ts-md5/dist/md5"
 import { useNavigate } from 'react-router-dom'
+
+import Faucet from 'components/faucet/Faucet'
 
 function TokenPage() {
 
@@ -45,11 +42,6 @@ function TokenPage() {
   const rootBalance = useSelector(selectlayer1Balance, isEqual)
   const layer = useSelector(selectLayer())
   const network = useSelector(selectNetwork())
-  const walletAddress = useSelector(selectWalletAddress())
-
-  const [ tweetUrl, setTweetUrl ] = useState("")
-  const [ isClaimFaucetLoading, setIsClaimFaucetLoading ] = useState(false)
-  const [ faucetErrorMsg, setFaucetErrorMsg ] = useState("")
 
   const [ debug, setDebug ] = useState(false)
 
@@ -61,17 +53,6 @@ function TokenPage() {
 
   const unorderedTransactions = useSelector(selectTransactions, isEqual)
   const orderedTransactions = orderBy(unorderedTransactions, i => i.timeStamp, 'desc')
-
-  let bobaTag = ''
-  if(walletAddress)
-    bobaTag = Md5.hashStr(walletAddress.toLowerCase().substring(2))
-
-  let BT = ''
-  let tweet = ''
-  if (bobaTag) {
-    BT = "BOBA" + bobaTag.substring(0, 9).toUpperCase()
-    tweet = "https://twitter.com/intent/tweet?text=I%27m%20developing%20on%20Boba%20Network%20" + BT
-  }
 
   const pendingL1 = orderedTransactions.filter((i) => {
     if (i.chain === 'L1pending' && //use the custom API watcher for fast data on pending L1->L2 TXs
@@ -127,11 +108,11 @@ function TokenPage() {
         return 'ethereum'
       } else if (i.symbolL1 === 'OMG') {
         return 'omg'
-      } else if(i.symbolL1 === 'BOBA') {
+      } else if (i.symbolL1 === 'BOBA') {
         return 'boba-network'
-      } else if(i.symbolL1 === 'OLO') {
+      } else if (i.symbolL1 === 'OLO') {
         return 'oolongswap'
-      } else if(i.symbolL1 === 'USDC') {
+      } else if (i.symbolL1 === 'USDC') {
         return 'usd-coin'
       } else {
         return i.symbolL1.toLowerCase()
@@ -147,31 +128,9 @@ function TokenPage() {
 
   const GasEstimateApprove = () => {
     let approval = networkService.estimateApprove()
-    console.log("GasEstimateApprove:",approval)
+    console.log(['Gas Estimate Approval', approval])
   }
 
-  async function claimAuthenticatedFaucetTokens() {
-    try {
-      setIsClaimFaucetLoading(true)
-      const tweetId = tweetUrl?.match(/twitter\.com\/.*\/status\/(\d+)/)[1]
-      const res = await networkService.getTestnetETHAuthenticatedMetaTransaction(tweetId)
-      if (!res) {
-        dispatch(openAlert('Faucet request submitted'))
-      } else {
-        setFaucetErrorMsg(res)
-      }
-    } catch (err) {
-      let error = err.message.match(/execution reverted: (.*)\\+"}}/)
-      if (error) {
-        error = error[1]
-      } else {
-        error = err?.message ?? err
-      }
-      setFaucetErrorMsg(error)
-    } finally {
-      setIsClaimFaucetLoading(false)
-    }
-  }
 
   if (!accountEnabled) {
 
@@ -199,176 +158,105 @@ function TokenPage() {
   } else {
 
     return (
-    <>
-      {layer === 'L2' && network === 'mainnet' &&
-        <Box sx={{ padding: '10px 0px', lineHeight: '0.9em' }}>
-          <Typography variant="body2">
-            <span style={{opacity: '0.9'}}>Need ETH or BOBA</span>{'? '}
-            <span style={{opacity: '0.6'}}>You can swap one for the other at </span>
-            <G.footerLink
-              target='_blank'
-              href={'https://www.sushi.com/swap'}
-              aria-label="link"
-              style={{fontSize: '1.0em', opacity: '0.9', paddingLeft: '3px'}}
-            >Sushiswap<Link />
-            </G.footerLink>
-            <span style={{opacity: '0.6'}}>and </span>
-            <G.footerLink
-              target='_blank'
-              href={'https://oolongswap.com/'}
-              aria-label="link"
-              style={{fontSize: '1.0em', opacity: '0.9', paddingLeft: '3px'}}
-            >Oolongswap <Link />
-            </G.footerLink>
-          </Typography>
-          {debug &&
-            <Button
-              onClick={()=>{GasEstimateApprove()}}
-              color='primary'
-              variant="contained"
-            >
-              GasEstimateApprove
-            </Button>
-          }
-        </Box>
-      }
-
-      {layer === 'L2' && network === 'goerli' &&
-          <G.LayerAlert style={{padding: '20px'}}>
-          <Box>
-
-            <Box style={{display: "inline-block"}}>
-              <Typography variant="body2">
-                Developer Twitter/Turing test token fountain - your Boba Bubble:{" "}
-                <span style={{ opacity: 0.65 }}>{BT} <Copy value={BT} light={false} /></span>
-              </Typography>
-            </Box>
-
-            <Typography variant="body3" sx={{ opacity: 0.65, marginBottom: "10px" }}>
-              Welcome developers.
-              For testnet BOBA and ETH, tweet your Boba Bubble and
-              then paste the tweet link in the field below.
+      <>
+        {layer === 'L2' && network === 'mainnet' &&
+          <Box sx={{ padding: '10px 0px', lineHeight: '0.9em' }}>
+            <Typography variant="body2">
+              <span style={{ opacity: '0.9' }}>Need ETH or BOBA</span>{'? '}
+              <span style={{ opacity: '0.6' }}>You can swap one for the other at </span>
+              <G.footerLink
+                target='_blank'
+                href={'https://www.sushi.com/swap'}
+                aria-label="link"
+                style={{ fontSize: '1.0em', opacity: '0.9', paddingLeft: '3px' }}
+              >Sushiswap<Link />
+              </G.footerLink>
+              <span style={{ opacity: '0.6' }}>and </span>
+              <G.footerLink
+                target='_blank'
+                href={'https://oolongswap.com/'}
+                aria-label="link"
+                style={{ fontSize: '1.0em', opacity: '0.9', paddingLeft: '3px' }}
+              >Oolongswap <Link />
+              </G.footerLink>
             </Typography>
-
-            <a
-              target='_blank'
-              rel="noopener noreferrer"
-              href={tweet}
-              aria-label="link"
-              style={{
-                backgroundColor: '#1b95e0',
-                color: '#fff',
-                borderRadius: '4px',
-                height: '28px',
-                fontWeight: '500',
-                fontSize: '13px',
-                lineheight: '26px',
-                padding: '8px 8px 8px 30px',
-                textDecoration: 'none',
-                backgroundImage: `url(${twitter})`,
-                backgroundRepeat: 'no-repeat',
-                backgroundSize: '16px 13px',
-                backgroundPosition: '8px 10px'
-              }}
-            >Tweet Now
-            </a>
-
-            <Typography variant="body3" sx={{ opacity: 0.65, marginTop: "10px", marginBottom: "10px" }}>
-              For the Tweet link, tap the share icon, tap "Share Tweet via", and finally select "Copy link to Tweet".
-            </Typography>
-
-            <Input
-              style={{width: '80%'}}
-              value={tweetUrl}
-              placeholder="Tweet Link"
-              onChange={(e) => setTweetUrl(e?.target?.value.split('?')[0])} //remove the superfluous stuff after the "?"
-            />
-
-            <Typography variant="body3" sx={{ opacity: 0.65, marginBottom: "10px", marginTop: '3px'}}>
-              You are limited to one fountain call per twitter account per day.
-              The transaction will not show in your history since it's a MetaTransaction (the gas is covered by Boba).
-            </Typography>
-
-            <Button
-              type="primary"
-              variant="contained"
-              style={{ marginTop: "10px", marginBottom: "18px" }}
-              disabled={!tweetUrl || !tweetUrl?.includes('http')}
-              loading={isClaimFaucetLoading}
-              onClick={async (e) => {await claimAuthenticatedFaucetTokens()}}
-              size="small"
-            >
-              Authenticated Faucet
-            </Button>
-
-            {faucetErrorMsg ? <Typography style={{color: 'red'}}>{faucetErrorMsg}</Typography> : null}
+            {debug &&
+              <Button
+                onClick={() => { GasEstimateApprove() }}
+                color='primary'
+                variant="contained"
+              >
+                GasEstimateApprove
+              </Button>
+            }
           </Box>
-          </G.LayerAlert>
-      }
+        }
 
-      {!!accountEnabled && inflight.length > 0 &&
-        <Box sx={{ padding: '10px 0px', display: 'flex', flexDirection: 'row' }}>
-          <Typography
-            variant="body2"
-            sx={{ cursor: 'pointer' }}
-            onClick={() => {
-              dispatch(setActiveHistoryTab("Pending"));
-              navigate('/history')
-            }}
-          >
-            <span style={{opacity: '0.9'}}>Bridge in progress:</span>{' '}
-            <span style={{opacity: '0.6'}}>Click for detailed status</span>
-            <Pulse variant="success"/>
-          </Typography>
-        </Box>
-      }
+        <Faucet />
 
-      <G.Container>
-        <G.Content>
-          <S.TableHeading>
-            {tokenTableHeads.map((item) => {
+        {!!accountEnabled && inflight.length > 0 &&
+          <Box sx={{ padding: '10px 0px', display: 'flex', flexDirection: 'row' }}>
+            <Typography
+              variant="body2"
+              sx={{ cursor: 'pointer' }}
+              onClick={() => {
+                dispatch(setActiveHistoryTab("Pending"));
+                navigate('/history')
+              }}
+            >
+              <span style={{ opacity: '0.9' }}>Bridge in progress:</span>{' '}
+              <span style={{ opacity: '0.6' }}>Click for detailed status</span>
+              <Pulse variant="success" />
+            </Typography>
+          </Box>
+        }
+
+        <G.Container>
+          <G.Content>
+            <S.TableHeading>
+              {tokenTableHeads.map((item) => {
+                return (
+                  <S.TableHeadingItem
+                    sx={{
+                      width: item.size,
+                      flex: item.flex,
+                      ...item.sx
+                    }}
+                    key={item.label} variant="body2" component="div">{item.label}</S.TableHeadingItem>
+                )
+              })}
+            </S.TableHeading>
+            {networkLayer === 'L2' ? !balanceLoading || !!childBalance.length ? childBalance.map((i, index) => {
               return (
-                <S.TableHeadingItem
-                  sx={{
-                    width: item.size,
-                    flex: item.flex,
-                    ...item.sx
-                  }}
-                  key={item.label} variant="body2" component="div">{item.label}</S.TableHeadingItem>
+                <ListToken
+                  key={i.currency}
+                  token={i}
+                  chain={'L2'}
+                  networkLayer={networkLayer}
+                  disabled={disabled}
+                />
               )
-            })}
-          </S.TableHeading>
-          {networkLayer === 'L2' ? !balanceLoading || !!childBalance.length ? childBalance.map((i, index) => {
-            return (
-              <ListToken
-                key={i.currency}
-                token={i}
-                chain={'L2'}
-                networkLayer={networkLayer}
-                disabled={disabled}
-              />
-            )
-          }) :
-          <S.LoaderContainer>
-            <CircularProgress color="secondary" />
-          </S.LoaderContainer> : null}
-          {networkLayer === 'L1' ? !balanceLoading || !!rootBalance.length ? rootBalance.map((i, index) => {
-            return (
-              <ListToken
-                key={i.currency}
-                token={i}
-                chain={'L1'}
-                networkLayer={networkLayer}
-                disabled={disabled}
-              />
-            )
-          }) :
-          <S.LoaderContainer>
-            <CircularProgress color="secondary" />
-          </S.LoaderContainer> : null}
-        </G.Content>
-      </G.Container>
-    </>)
+            }) :
+              <S.LoaderContainer>
+                <CircularProgress color="secondary" />
+              </S.LoaderContainer> : null}
+            {networkLayer === 'L1' ? !balanceLoading || !!rootBalance.length ? rootBalance.map((i, index) => {
+              return (
+                <ListToken
+                  key={i.currency}
+                  token={i}
+                  chain={'L1'}
+                  networkLayer={networkLayer}
+                  disabled={disabled}
+                />
+              )
+            }) :
+              <S.LoaderContainer>
+                <CircularProgress color="secondary" />
+              </S.LoaderContainer> : null}
+          </G.Content>
+        </G.Container>
+      </>)
   }
 
 }
